feat: filter chart values
