+++ src/pages/vebal.vue
@@ -2,7 +2,6 @@
 /**
  * veBAL page
  */
-import Hero from '@/components/contextual/pages/vebal/Hero.vue';
 import LMVoting from '@/components/contextual/pages/vebal/LMVoting/LMVoting.vue';
 import MyVeBAL from '@/components/contextual/pages/vebal/MyVeBAL/MyVeBAL.vue';
 import CrossChainBoostCards from '@/components/contextual/pages/vebal/cross-chain-boost/CrossChainBoostCards.vue';
@@ -11,6 +10,7 @@ import { isVeBalSupported } from '@/composables/useVeBAL';
 import { provideUserStaking } from '@/providers/local/user-staking.provider';
 import { providerUserPools } from '@/providers/local/user-pools.provider';
 import { providePoolStaking } from '@/providers/local/pool-staking.provider';
+import MyVebalInfo from '@/components/contextual/pages/vebal/MyVebalInfo.vue';
 
 const userStaking = provideUserStaking();
 providerUserPools(userStaking);
@@ -19,7 +19,7 @@ providePoolStaking();
 
 <template>
   <div>
-    <Hero v-if="isVeBalSupported" v-once />
+    <MyVebalInfo v-if="isVeBalSupported" v-once />
 
     <div class="py-16 xl:py-20 bg-gray-50 dark:bg-gray-850/50">
       <div v-if="isVeBalSupported" class="lg:container lg:mx-auto">
